move git exec module tests to pytest
